+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.5.2
 ### FEATURES
 - add key column to priorities and statuses tables (hra-done)
+- add seed script to populate database with system statuses and priorities
 
 ## 0.5.0
 ### FEATURES